Fix https://github.com/2d-inc/Flare-Flutter/issues/75
Fix similar to https://github.com/2d-inc/Flare-Flutter/issues/75

+++ lib/nima.dart
@@ -17,7 +17,7 @@ class FlutterActorImage extends ActorImage {
   Float32List _uvBuffer;
   ui.Paint _paint;
   ui.Vertices _canvasVertices;
-  Int32List _indices;
+  Uint16List _indices;
   DrawCallback onDraw;
 
   final Float64List _identityMatrix = Float64List.fromList(<double>[
@@ -63,7 +63,7 @@ class FlutterActorImage extends ActorImage {
     _vertexBuffer = makeVertexPositionBuffer();
     _uvBuffer = makeVertexUVBuffer();
     _indices =
-        Int32List.fromList(triangles); // nima runtime loads 16 bit indices
+        Uint16List.fromList(triangles); // nima runtime loads 16 bit indices
     updateVertexUVBuffer(_uvBuffer);
     int count = vertexCount;
     int idx = 0;
